+++ .size-limit.js
@@ -7,28 +7,28 @@ module.exports = [
 	{
 		name: 'i18n-object',
 		path: 'dist/i18n.object.min.js',
-		limit: '856 b',
+		limit: '892 b',
 	},
 	{
 		name: 'i18n-instance',
 		path: 'dist/i18n.min.js',
-		limit: '960 b',
+		limit: '994 b',
 	},
 	{
 		name: 'all together',
 		path: 'dist/i18n.all.min.js',
-		limit: '1013 b',
+		limit: '1046 b',
 	},
 	{
 		name: 'svelte-store',
 		path: 'svelte/svelte-store.min.js',
-		limit: '1104 b',
+		limit: '1147 b',
 		ignore: ['svelte'],
 	},
 	{
 		name: 'react-context',
 		path: 'react/react-context.min.js',
-		limit: '1080 b',
+		limit: '1124 b',
 		ignore: ['react'],
 	},
 ]